* commit '966eb002ba469a4e66e7c5041caa04689768eef3':
  fix type error for rpm

+++ can_sim/speed_p.py
@@ -44,7 +44,7 @@ class LimitSpeedP(Processer):
         super().__init__(0x118)
 
     def set_rpm(self, rpm):
-        self.set_data_auto(1, rpm / 4)
+        self.set_data_auto(1, int(rpm / 4))
 
     def set_rpm_valid(self, valid):
         self.set_data_bit_auto(2, 0, valid)
